type contrib/zip/lib .ch class declarations for M1 dispatch; fix K&R/rock/live bugs surfaced by typing
FossilOrigin-Name: 51a06c96580c4b48c5cf2ee8ba429f9945da78e2bcf1ba0529ad2de8784459c2

+++ src/contrib/zip/lib/zipedit.c
@@ -1257,7 +1257,7 @@ Pending_Delete( self, pane )
   zipedit_Cancel_Enclosure( self, pane );
   zipview_Draw_Pane( View, pane );
   zipedit_Hide_Selection_Menu( self );
-  zip_NotifyObservers( Data, View );
+  zip_NotifyObservers( Data, (long) View );
   OUT(Pending_Delete);
   }
